requirements/dev.txt
pillow
pillow_heif
pillow-avif-plugin